Add stg to git options
I've used stg enough to know that I love the workflow, and would like
to share it with the "wider chimera userbase". Therefore, making it an
enable option on git makes sense, as a tool to augment your git
experience. It may be worth doing the same for other tools like
gitbutler (as soon as there is a gitbutler nixpkg, which is expected
this weekend)

Change-Id: Ie763981becd93cf345f19f5c2e63f0566dbfd083
Reviewed-on: https://git.clicks.codes/c/Chimera/NixFiles/+/396
Reviewed-by: Samuel Shuert <coded@clicks.codes>
Tested-by: Skyler Grey <minion@clicks.codes>