Add mako
Mako is a notification daemon, and we /really/ need a notification
daemon if we're using hyprland (or programs like Firefox/Chrome will
create their own notifications which become full windows and mess up our
layout). This is a stopgap solution; probably we won't eventually go
with Mako because we don't think it meets our needs re: a place to see
past notifications, but for now it's far better than nothing.

As there are no alternatives in Chimera and it's essential to have a
good experience on hyprland, I have enabled it by default if hyprland is
enabled.

Change-Id: I581df76f0352a30fb8fcab74e7091089934b98be
Reviewed-on: https://git.clicks.codes/c/Chimera/NixFiles/+/430
Tested-by: Skyler Grey <minion@clicks.codes>
Reviewed-by: Samuel Shuert <coded@clicks.codes>