Add registry entries for all our inputs
I don't know most of what's in the nix registry (except nixpkgs and
pridefetch), but it would be nice to! This commit adds registry entries
for everything we care about and a lot besides.

Particularly of interest are nixpkgs, which is pinned now, and
templates, which points to our own /+/Templates repository where we
can fulfill our lifelong ambitions of having a standard gitignore for
every new repository!

Co-Authored-By: Skyler Grey <minion@clicks.codes>
But-I-Think: You could probably tell that
Change-Id: Ic70738906954638cf61bdb7c2766c214324738d9
Reviewed-on: https://git.clicks.codes/c/Chimera/NixFiles/+/417
Reviewed-by: Skyler Grey <minion@clicks.codes>
Tested-by: Samuel Shuert <coded@clicks.codes>