minion@greylag, coded@shorthair change password
Originally we were using initialPassword in modules/nixos/users
to set our passwords. Now we are setting them with hashed password
files. The files are stored in respective system directories and use
sops to encrypt them in place.

Finally, for this to be set for anything but the inital account creation
we need to disable mutableUsers (which is on by default), so let's do
that too.

Change-Id: I52c3658e7318d7d0c4ce0156582b754ea0b337c6
Reviewed-on: https://git.clicks.codes/c/Chimera/NixFiles/+/424
Reviewed-by: Samuel Shuert <coded@clicks.codes>
Tested-by: Skyler Grey <minion@clicks.codes>
Reviewed-by: Skyler Grey <minion@clicks.codes>