Add nix-index-database and comma
Nix-index-database is a premade database for nix-index. The process of
downloading a database is often rather slow, so building one into the
system config automatically makes sense

Comma is a command which integrates with nix-index to autorun commands
with the ability to detect what package a command is in. Using it allows
you to more easily run commands without having to go to search.nixos.org
and use nix shell or nix run

Change-Id: Ib44b65294152354a9ca433ad2315ddb2fcd8caa4
Reviewed-on: https://git.clicks.codes/c/Chimera/NixFiles/+/427
Tested-by: Skyler Grey <minion@clicks.codes>
Reviewed-by: Skyler Grey <minion@clicks.codes>