minion@greylag: add nextcloud-client, remove bat
Nextcloud-client is used to sync a folder with nextcloud. I sync my
notes to the Clicks nextcloud instance so not having the client means I
am unable to read/write my notes

Bat looks nice, but it has severe disadvantages to copypaste, and
doesn't display text at all in the (light) theme I'm using. All of these
issues are rectifiable, but it's easier to disable bat for now than fix
them

Change-Id: I5935470244bfe5d57c6e442d1f5b62b7794019e1
Reviewed-on: https://git.clicks.codes/c/Chimera/NixFiles/+/435
Tested-by: Skyler Grey <minion@clicks.codes>
Reviewed-by: Skyler Grey <minion@clicks.codes>