greylag: Silence boot logs
Boot prints out lots of text (e.g. graphics card warnings), all of which
is pointless to me or at the very least non-actionable. I also feel that
NixOS's rollbacking is good enough for me to be confident in disabling
boot logs for greylag.

I have not done so on a chimera level, as I feel that it's a personal
preference which you need to be quite aware of the risks to do. I feel
comfortable disabling logs on greylag but that doesn't mean that I have
them disabled on canvasback or that I am endorsing you to do the same.

Change-Id: Ie93987c66080ee2e955bf5e9e12b47b4c391c371
Reviewed-on: https://git.clicks.codes/c/Chimera/NixFiles/+/437
Tested-by: Skyler Grey <minion@clicks.codes>
Reviewed-by: Skyler Grey <minion@clicks.codes>