firefox: Add search engine options
I didn't default-enable bookmarks because enabling any home-manager
bookmark options will wipe your imperative bookmarks, so it's important
to make sure people know what they're getting into.

I also didn't default-enable extensions because I wanted to avoid
default-installing the DuckDuckGo privacy tools extension when this
feature was turned on, and because they are not necessary to use the
core search feature.

The images for NixOS searches are also too similar, it might be nice to
get some custom icons so we can tell at a glance which is is which

I considered adding NUR, but was unable to as they don't provide a
search page. This may be changed in this GitHub issue:
- https://github.com/nix-community/nur-search/issues/9

Change-Id: I70b54e2d98ecb89471ad5075518f67cc79038618
Reviewed-on: https://git.clicks.codes/c/Chimera/NixFiles/+/461
Tested-by: Skyler Grey <minion@clicks.codes>
Reviewed-by: Samuel Shuert <coded@clicks.codes>