Set git config pull.rebase to merges
We like rebases here, enough that we like having pull.rebase set by
default everywhere we go. Setting pull.rebase to merges differs from
plain pull.rebase = true in that it also rebases over merge commits so
they are linear. As we never want merge commits, this is a good default
for us

Change-Id: Ieba1a69aa5a6d0793d6ae4c490ba4d48f2917a09
Reviewed-on: https://git.clicks.codes/c/Chimera/NixFiles/+/514
Reviewed-by: Samuel Shuert <coded@clicks.codes>
Tested-by: Skyler Grey <minion@clicks.codes>