Use kitty-ssh if kitty is enabled
Kitty ssh is great for the vast majority of servers. There are 2
instances when it isn't great. Here they are:

- When you are not using kitty. The ssh kitten won't work at all in non-
  kitty terminals.
- When you are using an ssh server that isn't a standard terminal.
  Examples are join@tilde.town and git.charm.sh.

If you are in either of these situations, you can use \ssh ... to bypass
the alias. We also don't set this alias unless kitty is enabled

We could additionally check that kitty is being used by making sure we
have TERM correctly set, but this patch is fine for now

$ echo $TERM
xterm-kitty

Change-Id: I140841b68067ef001f2feea48b80c4f48f6be864
Reviewed-on: https://git.clicks.codes/c/Chimera/NixFiles/+/516
Tested-by: Skyler Grey <minion@clicks.codes>
Reviewed-by: Samuel Shuert <coded@clicks.codes>