greylag: add gnome-keyring
gnome-keyring is used to store passwords, and a large amount of programs
(e.g. Nextcloud sync client, etc.) rely on it existing.

Currently I've only enabled it for greylag, as this is a faster patch to
make. In future, I think we should enable this for everyone by default
(but we must not force enable it for everyone because alternative secret
providers such as KeePassXC exist)

Change-Id: Ic1d6abdc69253a158efb21055f7807251ac9116c
Reviewed-on: https://git.clicks.codes/c/Chimera/NixFiles/+/534
Reviewed-by: Skyler Grey <minion@clicks.codes>
Tested-by: Skyler Grey <minion@clicks.codes>