Add waybar
As eww was taking too long, we added a waybar configuration as a stopgap

Most of our inspiration was taken from
https://github.com/Andrey0189/hyprland-rice/tree/main/config/waybar,
with Nix integration from
https://github.com/Minion3665/NixFiles/blob/a124e5f0ac92ec39b22c469d6d85efbac018c28d/modules/waybar.nix

We have loosely rethemed this around the lack of several modules in the
title bar, as well as the need to make both a light and a dark theme,
but we still consider these close enough to the Andrey0189/hyprland-rice
code that we have licensed these files under MIT

We have defaulted this to be enabled when you enable hyprland, but this
may change in the future. To keep it when it does, please explicitly
enable it in your home configuration.

Change-Id: I64ffd1d2c9ebd008760a794f0bb7d1e81d4d3f32
Reviewed-on: https://git.clicks.codes/c/Chimera/NixFiles/+/569
Reviewed-by: Skyler Grey <minion@clicks.codes>
Tested-by: Samuel Shuert <coded@clicks.codes>