feat(greylag): limit CPU speed by default
The vast majority of the time, I don't need my CPU to go as fast as it
can. Unfortunately, it going as fast as it can will make it too hot so it
is unable to maintain it for prolonged periods of time.

Limiting it by default will prevent CPU heating while providing no
appreciable performance impact for regular activity.

Change-Id: I604835ffd4f096f5075b1746db71c69f1e225e70
Reviewed-on: https://git.clicks.codes/c/Chimera/NixFiles/+/701
Reviewed-by: Skyler Grey <minion@clicks.codes>
Tested-by: Skyler Grey <minion@clicks.codes>