fix(hyprland): use nixpkgs hyprland
Previously we've been using the hyprland flake. Unfortunately this has a
poor caching and stability story, particularly when different versions of
packages such as mesa are available in the nixpkgs we are using and the
nixpkgs hyprland flake.

This isn't a breaking change, because we are not reverting version
updates

Change-Id: I777f668af9da9ef7a51a4600283c71aa5a28a845
Reviewed-on: https://git.clicks.codes/c/Chimera/NixFiles/+/711
Tested-by: Skyler Grey <minion@clicks.codes>
Reviewed-by: Samuel Shuert <coded@clicks.codes>