feat(jj)!: Rebase ontop of master
Previously we were sticking to jj 18 due to the gerrit patch's version.
We've finally got around to rebasing that and pushed it up.

This isn't as good as solving the review comments - which probably needs
a bit more work but could take off any burden on us to continually
rebase this.

Change-Id: Ic5ebfb1713e94851a6c4293f9cf1bd33b1cb2a3c
Reviewed-on: https://git.clicks.codes/c/Chimera/NixFiles/+/822
Tested-by: Skyler Grey <minion@clicks.codes>
Reviewed-by: Samuel Shuert <coded@clicks.codes>