feat(git): Add jj
jj is a frontend for git, as well as a future version control system.
I've heard many good things, including that it is both simpler and more
powerful than git so I wanted to give it a test. Eventually I will
likely split this out into its own file as I suspect jj config will be
required, however for now this option is enough

Change-Id: Idc26653c2bcae71b0e8b0596bcbb466cafe64a6d
Reviewed-on: https://git.clicks.codes/c/Chimera/NixFiles/+/772
Reviewed-by: Samuel Shuert <coded@clicks.codes>
Tested-by: Skyler Grey <minion@clicks.codes>