build: create NixOS build workflow
I want to auto-update to my nixos config, but I don't want to build
everything in a million places! Therefore, I can push to a cache
and only download from a branch that I know is already cached

I need to build every NixOS system so I can cache them all, the easiest
way to do this is to just make a package that symlinks them all into
place

This package will need to be built once for every architecture we
support. Right now that's only x86_64. If we add home/darwin support we
will need separate packages for those