feat: retry failed auto-updates
It's fairly common for my laptop to fail auto-updating as it is off the
network. We can improve its chances by retrying auto-updates that fail,
and backing off if we're continuously failing. I pulled some random
numbers that look plausible for min/max/etc. delay: 5 seconds to 1 day
with some relatively small scaling steps. Since as this service will
also already be started once a day, I don't expect to ever reach the
high end (though there was some discussion about 1/week on servers(?))

Additionally, I've made the network wait until it is online using
networkmanager before it updates, since as the network being offline at
the start is by-far the largest reason for this to fail by my
observation (much more than config failures or dropping in the middle)

We should also set up notification of failed updates as soon as we have
a notification daemon