feat: make quickshell smaller on vertical monitors
Quickshell is currently too big for us to really add other widgets, so
it would make more sense to shrink it - here we use something that's
akin to css' vmin to decide how tall the area should be.

We decided to anchor the area to the top of the window (rather than the
center of the available area). Without doing so the bar looks weird when
we have the top workspace focused (i.e. a lot of the time). It's also
meant to be some sort of a bar-like thing (so anchoring it makes sense).