feat(minion): rebind capslock to ctrl/escape
On my [other keyboard][polyglot] I have a key that is both ctrl+escape
and it is consistently useful. It's also in a similar place to the caps
on a regular keyboard.

It would be nice, therefore, to have this key on any keyboard used with
my computers too. I considered rebinding with the caps:escape option in
Niri but this would not allow me to hold for control. Therefore, keyd
comes to the rescue!

[polyglot]: https://stenokeyboards.com/products/polyglot-keyboard