fix(niri): start ssh-agent later
Since switching niri to a systemd service we have gained some problems
with ssh-agent: it needs a restart to work.

We can't trivially pull niri later since gnome-keyring ssh wants to be
started as part of graphical-session-pre and depends on ssh-agent. To
fix this we can disable the gnome-keyring ssh module.