fix(copyparty): fix kanidm healthcheck
Previously we were using some code from a previous incarnation of OAuth2
Proxy (from Clicks) to determine if kanidm was started.

Unfortunately this missed that oauth2_proxy was now known as
oauth2-proxy so the health check was never getting activated.

Had it been activated we would also have seen that the default location
is always a redirect. Instead, it's better to use the official status
endpoint and check if it returns "true" (for "ready")