feat(common): move core packages from personal
We consider git and helix to be core packages: an editor which is useful
for making changes on the fly and git which is a core part of nix.

We have previously only had these installed in personal as they were in
configuration.nix at the start. This was an oversight. They should be in
common. Let's move them there.