fix(niri): start ssh agent after niri
Previously we were having the ssh-agent after graphical-session.target,
which sometimes seemed to work but mostly seemed to have dependency
ordering issues. In a bid to fix this, we set up some systemd-keyring
hacks (though seemed to fall short of actually setting .enable,
presumably because of the intermittence of this problem) and were about
to mask out the still-active systemd ssh keyring.

This, of course, would've been a mistake since as none of that was
actually causing the issue.

Instead, the problem was that niri seems to start after
graphical-session.target, we wanted to start after niri and were
wantedBy graphical-session.target

No good, obviously.

We can also get rid of the inactive gnome-keyring code - it's doing
nobody any good