feat: add scriptfs
ScriptFS is a filesystem that can run commands and use their output as
the virtual contents of files

We're using a fork-of-a-fork since as it supports fuse3 and built nicely
on Nix (in comparison to the source repo). The code is pretty small,
and generally a thing that I suspect would've been a "cool hack" to
get around some overly-limited program more than once. I also wonder if
we could use it effectively with Bitwarden to form part of our secrets
stack - i.e. in lieu of reading out secrets to a /run directory at the
start of day

We shouldn't run this as root since as it runs commands as the user who
started the filesystem - that's also why I avoided putting
`-o allow_other` on this one...

We're about to use this to do some unspeakable things to the signing
key path in Jujutsu - since as we want it dynamically determined by
what is in the SSH agent - and therefore we also need a symlink back to
/run/user/$UID from /home/$USER