feat(stalwart): add managesieve listener
Managesieve is used for users to manage sieve scripts: bits of code that
decide what should happen to an email when it is recieved. They can do
things like discarding mail from an address, filing mail in a specific
folder, etc.

We need to open up the port without implicit TLS as sieve clients seem
not to understand that...