feat(ingredients): write auto-imports for systems
We intend to switch the manual module importing to a system we call
"ingredients", where you will write a list of the modules you want
enabled and they will be enabled, along with any dependencies, etc.
without writing lots of .enable boilerplate by hand

In the long-term, this will allow us the following:
- Dependencies of some ingredients on other ingredients, rather than
  having to know and add all of them to your system manually
- Documentation generation of ingredient dependencies, etc.
- Combination ingredients (e.g. freshlybakedcake+personal, to be
  auto-imported if freshlybakedcake *and* personal both are)
- Auto-importing rather than having manual default.nixes on each level
  which we have previously forgotten to update

This is the start of that idea - it's not feature-complete but it does
auto-import files directly in ingredient subdirectories (it's unclear
if we want this to remain direct children only) and have a module with
enable arguments rather than direct imports. In other words: it's about
as minimal of a prototype as I could reasonably get

At the moment, I've only done system ingredients as, due to homes using
_module.args some of them need special provision for imports - I will
figure this out later (possibly we could specially handle the 'project'
argument for ingredients or have some method of overriding ingredient
import arguments?)