feat(systems): inherit home ingredients
We often have homes that have ingredients which need system components.
Examples are gaming (which needs steam to be installed globally),
nightshift (which needs geoclue2 to be installed globally), niri (which
needs some substitutors, xdg stuff, etc.). It's annoying to keep track
of these ourselves: far better to have any same-named ingredients be
added to attached systems when they are added to homes.

It's worth doing this in a module rather than pulling the ingredients
in to our ingredients list as otherwise we won't get ingredients enabled
as dependencies/etc.